Move WGPU engine to its own module
Disentangle the wgpu implementation of Engine from the mechanisms to create recordings, as the former can in theory be used with multiple back-ends.

Also change the name to WgpuEngine to emphasize that it is one choice among possibly many. Most clients are currently using the `Renderer` abstraction, which is not changed, so the refactor is fairly well hidden.
