+++ src/shaders.rs
@@ -26,10 +26,9 @@ use wgpu::Device;
 #[cfg(feature = "wgpu")]
 use crate::{
     cpu_shader,
-    engine::{BindType, Engine, Error, ImageFormat},
+    engine::{BindType, Error, ImageFormat, ShaderId},
+    wgpu_engine::WgpuEngine,
 };
-
-use crate::engine::ShaderId;
 
 macro_rules! shader {
     ($name:expr) => {&{
@@ -81,7 +80,7 @@ pub struct FullShaders {
 #[cfg(feature = "wgpu")]
 pub fn full_shaders(
     device: &Device,
-    engine: &mut Engine,
+    engine: &mut WgpuEngine,
     use_cpu: bool,
 ) -> Result<FullShaders, Error> {
     let imports = SHARED_SHADERS
